Pull request #4088: Build
Merge in SNORT/snort3 from ~RDEMPSTE/snort3:build to master

Squashed commit of the following:

commit 081a9a49a764b4e93b16047fa66592d3725dcb64
Author: Ron Dempster (rdempste) <rdempste@cisco.com>
Date:   Fri Nov 3 13:16:32 2023 -0400

    build: fix up 32-bit compilation

commit c34089572ac99037baa69003971fe9b8cfea5c2d
Author: Ron Dempster (rdempste) <rdempste@cisco.com>
Date:   Fri Nov 3 12:52:55 2023 -0400

    appid: clean up main thread appid debug and make appid on, off, on work

+++ src/memory/test/memory_cap_test.cc
@@ -235,7 +235,7 @@ TEST(memory, prune1)
     const uint64_t start = 50;
     heap->total = start;
 
-    MemoryConfig config { cap, 100, 0, 2, true };
+    MemoryConfig config { (size_t)cap, 100, 0, 2, true };
     MemoryCap::start(config, pruner);
     MemoryCap::thread_init();
 
@@ -336,7 +336,7 @@ TEST(memory, prune1)
 TEST(memory, prune3)
 {
     uint64_t cap = 100;
-    MemoryConfig config { cap, 100, 0, 1, true };
+    MemoryConfig config { (size_t)cap, 100, 0, 1, true };
     MemoryCap::start(config, pruner);
     MemoryCap::thread_init();
 
@@ -378,7 +378,7 @@ TEST(memory, prune3)
 TEST(memory, two_cycles)
 {
     uint64_t cap = 100;
-    MemoryConfig config { cap, 100, 0, 1, true };
+    MemoryConfig config { (size_t)cap, 100, 0, 1, true };
     MemoryCap::start(config, pruner);
     MemoryCap::thread_init();
 
@@ -423,7 +423,7 @@ TEST(memory, reap_failure)
     const uint64_t start = 50;
     heap->total = start;
 
-    MemoryConfig config { cap, 100, 0, 2, true };
+    MemoryConfig config { (size_t)cap, 100, 0, 2, true };
     MemoryCap::start(config, pruner);
     MemoryCap::thread_init();
 
@@ -481,7 +481,7 @@ TEST(memory, reap_freed_outside_of_pruning)
     const uint64_t start = 50;
     heap->total = start;
 
-    MemoryConfig config { cap, 100, 0, 2, true };
+    MemoryConfig config { (size_t)cap, 100, 0, 2, true };
     MemoryCap::start(config, pruner);
     MemoryCap::thread_init();
